Add support for specifying an operation name
This adds Graphql.Http.withOperationName so you can write something
like:

```
makeRequest : Cmd Msg
makeRequest =
    query
        |> Graphql.Http.queryRequest "https://elm-graphql.herokuapp.com"
        |> Graphql.Http.withOperationName "SayHello"
        |> Graphql.Http.send (RemoteData.fromResult >> GotResponse)
```

The operation name is specified in the GrpahQL docs:
https://graphql.org/learn/queries/#operation-name

The operation name is a meaningful and explicit name for your operation.
It is only required in multi-operation documents, but its use is encouraged
because it is very helpful for debugging and server-side logging.