Add support for the "extensions" key on errors
The latest (published) graphql spec standardized an "extensions" key in
the response format of graphql errors for optionally sending additional
information about the error. The value of the extensions key is required
to be a map, but beyond that there is no additional specification.

This PR adds support for graphql error extensions to the
Graphql.Http.GraphqlError.GraphqlError record in the form of a `Maybe
(Dict String Json.Decode.Value)` field, allowing users to handle the
values however they please.